+++ locales/en_US/index.mjs
@@ -1690,5 +1690,9 @@ export default {
   "expression_func_display_value_p0_value": "value",
   "expression_func_display_value_p0_description": "A reference to an element holding the value.",
   "expression_func_display_value_p1_value": "'full_path'",
-  "expression_func_display_value_p1_description": "The <b>full path</b> (not data path) of the field between single quotes, eg. `DISPLAY_VALUE(container.select, 'container.select')`. Siblings path (eg. 'list.*.select') cannot be used."
+  "expression_func_display_value_p1_description": "The <b>full path</b> (not data path) of the field between single quotes, eg. `DISPLAY_VALUE(container.select, 'container.select')`. Siblings path (eg. `list.*.select`) cannot be used.",
+  "expression_func_available_description": "Returns `true` if the element is available (all of its conditions are met).",
+  "expression_func_available_p0_value": "'full_path'",
+  "expression_func_available_p0_description": "The <b>full path</b> (not data path) of the field between single quotes, eg. `AVAILABLE('container.select')`. Siblings path (eg. `list.*.select`) cannot be used.",
+	"rename_error_number_start": "Must not start with number"
 };